#3: changed package and project structure after CR
