--- src/main/java/net/mySmirnov/quiz/ui/InputUI.java
@@ -1,5 +1,0 @@
-package net.mySmirnov.quiz.ui;
-
-public interface InputUI {
-    String read();
-}
