--- src/main/java/net/mySmirnov/quiz/ui/OutputUI.java
@@ -1,5 +1,0 @@
-package net.mySmirnov.quiz.ui;
-
-public interface OutputUI {
-    void write(Object s);
-}
