--- src/main/java/net/mySmirnov/quiz/ui/OutputUIImpl.java
@@ -1,8 +1,0 @@
-package net.mySmirnov.quiz.ui;
-
-public class OutputUIImpl implements OutputUI {
-    @Override
-    public void write(Object s) {
-        System.out.println(s);
-    }
-}
