+++ src/main/java/net/mysmirnov/quiz/QuizApplication.java
@@ -1,0 +1,8 @@
+package net.mysmirnov.quiz;
+
+public interface QuizApplication {
+    /**
+     * Этот метод запускае приложение
+     */
+    void run();
+}
