+++ src/main/java/net/mysmirnov/quiz/ui/InputUIService.java
@@ -1,0 +1,5 @@
+package net.mysmirnov.quiz.ui;
+
+public interface InputUIService {
+    String read();
+}
