+++ src/main/java/net/mysmirnov/quiz/ui/OutputUIService.java
@@ -1,0 +1,5 @@
+package net.mysmirnov.quiz.ui;
+
+public interface OutputUIService {
+    void write(Object s);
+}
