+++ src/main/java/net/mysmirnov/quiz/ui/OutputUIServiceImpl.java
@@ -1,0 +1,8 @@
+package net.mysmirnov.quiz.ui;
+
+public class OutputUIServiceImpl implements OutputUIService {
+    @Override
+    public void write(Object s) {
+        System.out.println(s);
+    }
+}
