+++ src/main/resources/context.xml
@@ -9,13 +9,19 @@
 
     <context:property-placeholder location="classpath:application.properties"/>
 
-    <bean id="quizApplication" class="net.mySmirnov.quiz.QuizApplication">
+    <bean id="quizApplication" class="net.mysmirnov.quiz.QuizApplicationImpl">
         <constructor-arg index="0" ref="questionService"/>
+        <constructor-arg index="1" ref="inputUIService"/>
+        <constructor-arg index="2" ref="outputUIService"/>
     </bean>
 
-    <bean id="questionService" class="net.mySmirnov.quiz.service.CsvQuestionService" init-method="init">
+    <bean id="questionService" class="net.mysmirnov.quiz.service.CsvQuestionService" init-method="init">
         <property name="maxNumberOfWrongAnswers" value="${max.number.of.wrong.answers}"/>
         <property name="resourceName" value="/data.csv"/>
     </bean>
+
+    <bean id="inputUIService" class="net.mysmirnov.quiz.ui.InputUIServiceImpl"/>
+
+    <bean id="outputUIService" class="net.mysmirnov.quiz.ui.OutputUIServiceImpl"/>
 
 </beans>
