+++ src/test/java/net/mysmirnov/quiz/service/CsvQuestionServiceTest.java
@@ -1,5 +1,6 @@
-package net.mySmirnov.quiz.service;
+package net.mysmirnov.quiz.service;
 
+import net.mysmirnov.quiz.service.CsvQuestionService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
